chore: code cleanup, rename files for to maintain consistency
