+++ src/components/FaqSection/styles.jsx
