+++ src/components/HeroSection/IconsList.js
