+++ src/components/HeroSection/TickerItem.jsx
@@ -1,4 +1,6 @@
 import React from 'react';
+
+// Libraries
 import styled from 'styled-components';
 import tw from 'twin.macro';
