+++ src/components/SponsorsSection/styles.jsx
